restructure executors
