+++ packages/nx-phrase/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@porscheofficial/nx-phrase",
-    "version": "0.0.5",
+    "version": "0.0.6",
     "repository": {
         "type": "url",
         "url": "https://github.com/porscheofficial/nx-plugins",
