--- packages/nx-phrase/src/executors/find-unused/schema.d.ts
@@ -1,6 +1,0 @@
-import { NonSensitiveArgs } from "../lib/types"
-
-export interface UnusedExecutorSchema extends Partial<NonSensitiveArgs> {
-    sourceKeyTransformer?: string
-    phraseKeyTransformer?: string
-} // eslint-disable-line
